Main functionality should be restored!
I have yet to fully test everything, but it seems like the main functionality works.

I also added in logging and I'm in the middle of improving the way that the --verbose option is implemented! Currently, logging does some weird stuff where it double prints some things, but it's not too big of an issue and it seems like there might be some fixes on SO! :)

Very exciting stuff.