Bug fixing and detail ironing! :O :D :)
Made a lot of small changes. I got rid of some extraneous code, updated README.md, and improved the way plots are exported.

I made a .ipynb to test the package, but the notebook did not want to run properly... I then made a regular .py script, and everything worked fine (then again, I was updating everything and not using it as a package, so we'll see how that goes! :)).

Using the .py script I made, I'll soon be updating the README.md with how to use CRICKETS as a Python package soon!