Odd changes, more digging needs to be done...
Some updates to README.md, but also some weird issues. It seems that when I run analysis with the plotting code, it runs twice after generating the plots and output table. When it does so, it flags a different number of bins as being dirty. I think the issue might be that somewhere the plotting functions are calling the get_exkurt function with a different set of parameters than the user input?